converted to cpp with templates
