+++ makefile
@@ -1,0 +1,3 @@
+./bin/yaep:
+	g++ ./src/yaep.cpp -o ./bin/yaep
+
